Moved scenes directory to src
