--- scenes/nodeRedTopics.js
@@ -1,2 +1,0 @@
-export const INIT_REST_NODE_CLIENT = 'INIT_REST_NODE_CLIENT'
-export const START_BEDTIME = 'START_BEDTIME'
