+++ src/scenes/nodeRedTopics.js
@@ -1,0 +1,2 @@
+export const INIT_REST_NODE_CLIENT = 'INIT_REST_NODE_CLIENT'
+export const START_BEDTIME = 'START_BEDTIME'
