+++ src/scenes/uiBuilderReactEvents.js
@@ -1,0 +1,12 @@
+// Please see the uibuilder documentation for the list of available events
+
+export const CONTROL_MESSAGE = 'ctrlMsg'
+export const IO_CONNECTED = 'ioConnected'
+export const ON_MESSAGE = 'msg'
+export const RECEIVED_CONTROL_MESSAGES = 'msgsCtrl'
+export const RECEIVED_MESSAGES_COUNT = 'msgsReceived'
+export const SENT_CONTROL_MESSAGE = 'sentCtrlMsg'
+export const SENT_CONTROL_MESSAGES_COUNT = 'msgsSentCtrl'
+export const SENT_MESSAGE = 'msgSent'
+export const SENT_MESSAGES_COUNT = 'msgsSent'
+export const SERVER_TIME_OFFSET = 'serverTimeOffset'
